refactor: application 패키지 분리
