+++ src/main/java/application/handler/resolver/ViewResolver.java
@@ -1,4 +1,4 @@
-package webserver.handler.resolver.view;
+package application.handler.resolver;
 
 import webserver.handler.resolver.AbstractResolver;
 import webserver.http.content.Content;
