+++ src/main/java/application/model/User.java
@@ -1,4 +1,4 @@
-package model;
+package application.model;
 
 public class User {
     private String userId;
