+++ src/main/java/webserver/utils/IOUtils.java
@@ -1,14 +1,12 @@
-package utils;
+package webserver.utils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 
 public class IOUtils {
     /**
-     * @param BufferedReader는
-     *            Request Body를 시작하는 시점이어야
-     * @param contentLength는
-     *            Request Header의 Content-Length 값이다.
+     * @param BufferedReader는 Request Body를 시작하는 시점이어야
+     * @param contentLength는  Request Header의 Content-Length 값이다.
      * @return
      * @throws IOException
      */
