+++ src/test/java/utils/IOUtilsTest.java
@@ -3,6 +3,7 @@ package utils;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.utils.IOUtils;
 
 import java.io.BufferedReader;
 import java.io.StringReader;
