--- src/test/java/webserver/request/path/PathVariableTest.java
@@ -1,6 +1,0 @@
-package webserver.request.path;
-
-class PathVariableTest {
-
-
-}
